Separate FAL poller
